refactor: move ACS validation API to SDK
Make acs-generator a versioned CLI-only package and move reusable validation, schemas, diagnostics, and tests into the Python SDK validation extra.


Copilot-Session: f5774aa9-5db0-4fe9-a483-b537b7460796
