+++ policy-engine/generator/acs_generator/schema/__init__.py
@@ -1,1 +1,4 @@
-"""Packaged ACS manifest schema."""
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+"""Packaged ACS generator wire schemas."""
