+++ policy-engine/sdk/python/agent_control_specification/schema/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 
-"""Command-line ACS policy artifact generator."""
+"""Packaged ACS validation schemas."""
